requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
pypdf
requests   
ollama
